Add: Send & Connect timeout, Max callback, gzip decompression field in schema
Add: Send & Connect timeout, Max callback, gzip decompression field in schema
Refactor: Change keeplive to 5sec and event queue size to 5K
Refactor: Update zzlib decompression code
Fix: Set blocked by field to nil when event is not blocked by governance rule
Refactor: Update README.md
