+++ README.md
@@ -151,7 +151,7 @@ Here's a list of all the parameters which can be used in this plugin's configura
 |config.application_id	||The Moesif application token provided to you by Moesif.|
 |config.api_endpoint|https://api.moesif.net|URL for the Moesif API.|
 |config.timeout|10000|An optional timeout in milliseconds when sending data to Moesif.|
-|config.keepalive|10000|An optional value in milliseconds that defines for how long an idle connection will live before being closed.|
+|config.keepalive|5000|An optional value in milliseconds that defines for how long an idle connection will live before being closed.|
 |config.api_version|1.0|An optional API Version you want to tag this request with in Moesif.|
 |config.disable_capture_request_body|false|An option to disable logging of request body.|
 |config.disable_capture_response_body|false|An option to disable logging of response body.|
