+++ kong-plugin-moesif-0.2.17-1.rockspec
@@ -37,6 +37,8 @@ build = {
     ["kong.plugins.moesif.connection"] = "kong/plugins/moesif/connection.lua",
     ["kong.plugins.moesif.lib_deflate"] = "kong/plugins/moesif/lib_deflate.lua",
     ["kong.plugins.moesif.client_ip"] = "kong/plugins/moesif/client_ip.lua",
+    ["kong.plugins.moesif.inflate-bit32"] = "kong/plugins/moesif/inflate-bit32.lua",
+    ["kong.plugins.moesif.inflate-bwo"] = "kong/plugins/moesif/inflate-bwo.lua",
     ["kong.plugins.moesif.zzlib"] = "kong/plugins/moesif/zzlib.lua",
     ["kong.plugins.moesif.base64"] = "kong/plugins/moesif/base64.lua"
   }
