+++ kong/plugins/moesif/connection.lua
@@ -16,7 +16,7 @@ function _M.get_connection(url_path, conf, sock)
   local host = parsed_url.host
   local port = tonumber(parsed_url.port)
 
-  sock:settimeout(conf.timeout)
+  sock:settimeout(conf.connect_timeout)
   local api_version = conf.api_version
   ngx.ctx.api_version = api_version
   local ok, err = sock:connect(host, port)
