+++ kong/plugins/moesif/schema.lua
@@ -2,8 +2,10 @@ return {
   fields = {
     api_endpoint = {required = true, type = "url", default = "https://api.moesif.net"},
     timeout = {default = 1000, type = "number"},
-    keepalive = {default = 2000, type = "number"},
-    event_queue_size = {default = 2000, type = "number"},
+    connect_timeout = {default = 1000, type = "number"},
+    send_timeout = {default = 2000, type = "number"},
+    keepalive = {default = 5000, type = "number"},
+    event_queue_size = {default = 5000, type = "number"},
     api_version = {default = "1.0", type = "string"},
     application_id = {required = true, default ="", type="string"},
     disable_capture_request_body = {default = false, type = "boolean"},
@@ -17,8 +19,9 @@ return {
     batch_size = {default = 200, type = "number"},
     disable_transaction_id = {default = false, type = "boolean"},
     debug = {default = false, type = "boolean"},
-    enable_skip_gzip_payload_decompression = {default = true, type = "boolean"},
+    disable_gzip_payload_decompression = {default = false, type = "boolean"},
     user_id_header = {default = "", type = "string"},
     company_id_header = {default = "", type = "string"},
+    max_callback_time_spent = {default = 2000, type = "number"},
   }
 }
